src/requirements.txt
pytesseract
pillow
pandas